Convert common.h helpers to NAPI
Convert BooleanOptionValue() and UInt32OptionValue() helpers to NAPI.

This change decides that the equivalent of v8::Local<>::IsEmpty is
napi_value == nullptr.

napi_env must be plumbed through to these helpers now.  napi_env will
have to be plumbed through to any function that wants to use NAPI
functions.  This may be a cumbersome burden on the user.  On the other
hand it is no different than v8's Isolate.  Nan avoids this by choosing
to use v8::Isolate::GetCurrent() everywhere, but this may not work well
if node.js allows multiple isolates to work with a single module.